+++ UFC/src/main/java/edu/kh/cooof/gym/trainerSelect/model/mapper/TrainerSelectMapper.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Mapper;
 
 import edu.kh.cooof.gym.trainerSelect.model.dto.PaymentRequest;
+import edu.kh.cooof.gym.trainerSelect.model.dto.PtPrice;
 import edu.kh.cooof.gym.trainerSelect.model.dto.Trainer;
 
 @Mapper
@@ -26,6 +27,9 @@ public interface TrainerSelectMapper {
 	 * @return
 	 */
 	int gymPayment(PaymentRequest request);
+	
+	// 로그인한 회원 pt 정보 조회
+		PtPrice selectPtPriceByMemberNo(int memberNo);
