+++ UFC/src/main/java/edu/kh/cooof/gym/trainerSelect/model/service/TrainerSelectService.java
@@ -3,6 +3,7 @@ package edu.kh.cooof.gym.trainerSelect.model.service;
 import java.util.List;
 
 import edu.kh.cooof.gym.trainerSelect.model.dto.PaymentRequest;
+import edu.kh.cooof.gym.trainerSelect.model.dto.PtPrice;
 import edu.kh.cooof.gym.trainerSelect.model.dto.Trainer;
 
 public interface TrainerSelectService {
@@ -26,6 +27,9 @@ public interface TrainerSelectService {
 	 */
 	int gymPayment(PaymentRequest request);
 
+	
+	// 로그인한 회원의 pt 정보
+	PtPrice getPriceByMemberNo(int memberNo);
